REF : ui 로직 분리
